Ajustado tabelas de departamento, craida função de desvincular membro do departamento e criada tabela de membro to departamento

+++ app/Http/Controllers/Dpt.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\Departamento;
 use App\Models\DepartamentoUsuario;
 use App\Models\User;
+use Illuminate\Support\Facades\DB;
 
 class Dpt extends Controller
 {
@@ -14,22 +15,24 @@ class Dpt extends Controller
             return back()->with('error', 'você precisa de permissão para fazer isso');
         }
 
-        $dpts = Departamento::where('nome', 'like', '%'.$request->busca.'%')->paginate(4);
-        //$dpts->toArray;
+        //Pegando todos os departamentos de pela busca
+        $dpts = Departamento::where('nome', 'like', '%'.$request->buscaDpt.'%')->paginate(4,['*'],'dptPage');
         //dd($dpts);
 
-
+        //adicionando a quantidade de pessoas vnculadas a cada departamento
         for($x = 0; $x<count($dpts);$x++){
             $dpts[$x]['qtdPessoa'] = $qtdPessoa = DepartamentoUsuario::where('departamento_id','=',$dpts[$x]->id)->count();
         }
         //dd($dpts);
 
-        $allDpt = Departamento::all();
+        //Pegando todos os usuarios
         $allUser = User::where('id', '!=', 1)->get();
-        //dd($allDpt);
+        //dd($allUser)
 
+        $pessoaDpt = DB::table('departamento_user as du')->join('users as u', 'u.id', '=', 'du.user_id')->join('departamentos as d', 'd.id', '=', 'du.departamento_id')->select('du.id','u.apelido as nome', 'd.nome as departamento')->where('u.apelido','like','%'.$request->buscaApelido.'%')->where('d.nome','like','%'.$request->buscaDpt2.'%')->paginate(4,['*','pessoaPage']);
 
-        return view('pages.dpt.dpt_home', compact('dpts','allDpt','allUser'));
+        //dd($pessoaDpt);
+        return view('pages.dpt.dpt_home', compact('dpts','allUser','pessoaDpt'));
     }
 
     public function adicionarPessoa(Request $request, string $idDpt){
@@ -59,6 +62,9 @@ class Dpt extends Controller
 
     public function criarDpt(Request $request){
         //dd($request);
+        if(auth()->user()->tipo != 'admin' && auth()->user()->tipo != 'pastor'){
+            return back()->with('error', 'você precisa de permissão para fazer isso');
+        }
 
         if(empty($request->nome)){
             return back()->with('error', 'É necessario escolher um nome para o Departamento');
@@ -77,10 +83,39 @@ class Dpt extends Controller
 
     public function excluirDpt(string $id){
         //dd($id);
+        if(auth()->user()->tipo != 'admin' && auth()->user()->tipo != 'pastor'){
+            return back()->with('error', 'você precisa de permissão para fazer isso');
+        }
 
+        $dptUser = DepartamentoUsuario::where('departamento_id',$id)->get();
+        //dd($dptUser);
+
+        foreach ($dptUser as $key){
+            $key->delete();
+        }
         $dpt = Departamento::find($id);
         $dpt->delete();
 
         return redirect()->route('dpt.index')->with('success', 'Departamento Excluido com sucesso');
+    }
+
+    public function desvincularpessoa(string $nome_pessoa, string $nome_dpt){
+        if(auth()->user()->tipo != 'admin' && auth()->user()->tipo != 'pastor' && auth()->user()->tipo != 'lider'){
+            return back()->with('error', 'você precisa de permissão para fazer isso');
+        }
+
+        $user = User::where('apelido',$nome_pessoa)->first();
+        $dpt = Departamento::where('nome',$nome_dpt)->first();
+        //dd($user,$dpt);
+
+        if(!$user || !$dpt){
+            return back()->with('error','Usuário ou Departamento não encontrados');
+        }
+
+        if(DepartamentoUsuario::where('user_id',$user->id)->where('departamento_id',$dpt->id)->exists()){
+            DepartamentoUsuario::where('user_id',$user->id)->where('departamento_id',$dpt->id)->delete();
+            return back()->with('success','Usuário desvinculado com sucesso');
+        }
+
     }
 }
